Mark LTS unavailable when giving 5oC reading
When TRVs lose signal the temperature sensors report a temp of 5oC even though the temperature hasn't changed. This can't be distinguished from a true reading in other components and distorts average temperature calculations when calculated from the helper or template sensors. Applying this change to the LTS sensor only will rectify this while not affecting the main sensor. However, the same could be applied to the regular climate sensors if thought beneficial.